Backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
supabase>=2.11.0
stripe>=11.4.0
PyJWT[crypto]>=2.9.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
httpx>=0.28.0
slowapi>=0.1.9
python-dotenv>=1.0.1
certifi>=2024.0.0
